Added ENV-friendly `database.yml` file.
One can set the DATABASE_URL variable, which will set up the whole
connection, or use one or more of:
- DATABASE_ADAPTER
- DATABASE_HOST
- DATABASE_PORT
- DATABASE_NAME
- DATABASE_USER
- DATABASE_PASSWORD

Each of them will just override a specific entry of the default
configuration, which is under the `"defaults"` key.

In a dev machine, one could probably just set DATABASE_USER and
DATABASE_PASSWORD. In a virtual server, the easiest way is to provide
DATABASE_URL. In heroku and similar providers, this file will be
overwritten.