Replace DATABASE_URL with DATABASE_NAME
We used to connect to it through the PostgreSQL URL but that ENV var is
no longer present in the server and we have `DATABASE_NAME` instead.
That's what we already used successfully with the `next` environment.

This is required to deploy using the `production` environment or the app
won't be able to connect to the DB.