Allow non-www host name
As far as I know our DNS provider redirects itself timeoverflow.org to
www.timeoverflow.org instead of our Nginx, as we are used to, but we're
seeing people getting ERR_CONNECTION_REFUSED, with mobile data only,
they say. We hope this may fix it but we're totally blind here.